move into new index module
move core modules to index module
