+++ src/index/segment_component.rs
