+++ src/indexer/single_segment_index_writer.rs
