+++ src/query/phrase_query/phrase_weight.rs
@@ -1,6 +1,6 @@
 use super::PhraseScorer;
-use crate::core::SegmentReader;
 use crate::fieldnorm::FieldNormReader;
+use crate::index::SegmentReader;
 use crate::postings::SegmentPostings;
 use crate::query::bm25::Bm25Weight;
 use crate::query::explanation::does_not_match;
