+++ src/query/query_parser/query_parser.rs
@@ -13,7 +13,7 @@ use super::logical_ast::*;
 use crate::core::json_utils::{
     convert_to_fast_value_and_get_term, set_string_and_get_terms, JsonTermWriter,
 };
-use crate::core::Index;
+use crate::index::Index;
 use crate::query::range_query::{is_type_valid_for_fastfield_range_query, RangeQuery};
 use crate::query::{
     AllQuery, BooleanQuery, BoostQuery, EmptyQuery, FuzzyTermQuery, Occur, PhrasePrefixQuery,
